+++ flutter_map/lib/src/layer/polyline_layer.dart
@@ -1,8 +1,10 @@
-import 'package:flutter/widgets.dart';
-import '../map/map.dart';
-import 'package:latlong/latlong.dart';
-import '../../flutter_map.dart';
 import 'dart:ui';
+
+import 'package:flutter/widgets.dart';
+import 'package:latlong/latlong.dart';
+
+import '../../flutter_map.dart';
+import '../map/map.dart';
 
 class PolylineLayerOptions extends LayerOptions {
   final List<Polyline> polylines;
@@ -27,6 +29,15 @@ class PolylineLayer extends StatelessWidget {
   PolylineLayer(this.polylineOpts, this.map);
 
   Widget build(BuildContext context) {
+    return new LayoutBuilder(
+      builder: (BuildContext context, BoxConstraints bc) {
+        final size = new Size(bc.maxWidth, bc.maxHeight);
+        return _build(context, size);
+      },
+    );
+  }
+
+  Widget _build(BuildContext context, Size size) {
     return new StreamBuilder<int>(
       stream: map.onMoved, // a Stream<int> or null
       builder: (BuildContext context, _) {
@@ -35,8 +46,7 @@ class PolylineLayer extends StatelessWidget {
           var i = 0;
           for (var point in polylineOpt.points) {
             var pos = map.project(point);
-            pos = pos.multiplyBy(map.getZoomScale(map.zoom, map.zoom)) -
-                map.getPixelOrigin();
+            pos = pos.multiplyBy(map.getZoomScale(map.zoom, map.zoom)) - map.getPixelOrigin();
             polylineOpt.offsets.add(new Offset(pos.x.toDouble(), pos.y.toDouble()));
             if (i > 0 && i < polylineOpt.points.length) {
               polylineOpt.offsets.add(new Offset(pos.x.toDouble(), pos.y.toDouble()));
@@ -50,6 +60,7 @@ class PolylineLayer extends StatelessWidget {
           polylines.add(
             new CustomPaint(
               painter: new PolylinePainter(polylineOpt),
+              size: size,
             ),
           );
         }
@@ -73,8 +84,11 @@ class PolylinePainter extends CustomPainter {
     if (polylineOpt.offsets.isEmpty) {
       return;
     }
-    var paint = new Paint()..color = polylineOpt.color;
-    paint.strokeWidth = polylineOpt.strokeWidth;
+    final rect = Offset.zero & size;
+    canvas.clipRect(rect);
+    final paint = new Paint()
+      ..color = polylineOpt.color
+      ..strokeWidth = polylineOpt.strokeWidth;
     canvas.drawPoints(PointMode.lines, polylineOpt.offsets, paint);
   }