+++ flutter_map/lib/src/map/map.dart
@@ -79,8 +79,20 @@ class MapState {
   }
 
   void move(LatLng center, double zoom) {
+    if (options.isOutOfBounds(center)) {
+      return;
+    }
+
     if (zoom == null) {
       zoom = _zoom;
+    }
+
+    // Abide to min/max zoom
+    if (options.maxZoom != null) {
+      zoom = (zoom > options.maxZoom) ? options.maxZoom : zoom;
+    }
+    if (options.minZoom != null) {
+      zoom = (zoom < options.minZoom) ? options.minZoom : zoom;
     }
 
     _zoom = zoom;
@@ -189,5 +201,13 @@ class MapState {
   Point getNewPixelOrigin(LatLng center, [double zoom]) {
     var viewHalf = this.size / 2.0;
     return (this.project(center, zoom) - viewHalf).round();
+  }
+
+  Bounds getPixelBounds(double zoom) {
+    var mapZoom = zoom;
+    var scale = getZoomScale(mapZoom, zoom);
+    var pixelCenter = project(center, zoom).floor();
+    Point<num> halfSize = size / (scale * 2);
+    return new Bounds(pixelCenter - halfSize, pixelCenter + halfSize);
   }
 }