+++ flutter_map_example/lib/widgets/drawer.dart
@@ -2,9 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_map_example/pages/esri.dart';
 import 'package:flutter_map_example/pages/home.dart';
 import 'package:flutter_map_example/pages/map_controller.dart';
+import 'package:flutter_map_example/pages/animated_map_controller.dart';
 import 'package:flutter_map_example/pages/marker_anchor.dart';
+import 'package:flutter_map_example/pages/plugin_api.dart';
 import 'package:flutter_map_example/pages/polyline.dart';
 import 'package:flutter_map_example/pages/tap_to_add.dart';
+import 'package:flutter_map_example/pages/offline_map.dart';
 
 Drawer buildDrawer(BuildContext context, String currentRoute) {
   return new Drawer(
@@ -51,10 +54,31 @@ Drawer buildDrawer(BuildContext context, String currentRoute) {
           },
         ),
         new ListTile(
+          title: const Text('Animated MapController'),
+          selected: currentRoute == AnimatedMapControllerPage.route,
+          onTap: () {
+            Navigator.popAndPushNamed(context, AnimatedMapControllerPage.route);
+          },
+        ),
+        new ListTile(
           title: const Text('Marker Anchors'),
           selected: currentRoute == MarkerAnchorPage.route,
           onTap: () {
             Navigator.popAndPushNamed(context, MarkerAnchorPage.route);
+          },
+        ),
+        new ListTile(
+          title: const Text('Plugins'),
+          selected: currentRoute == PluginPage.route,
+          onTap: () {
+            Navigator.popAndPushNamed(context, PluginPage.route);
+          },
+        ),
+        new ListTile(
+          title: const Text('Offline Map'),
+          selected: currentRoute == OfflineMapPage.route,
+          onTap: () {
+            Navigator.popAndPushNamed(context, OfflineMapPage.route);
           },
         ),
       ],